requirements.txt
streamlit
numpy
Pillow  # pour PIL
pandas
sentence-transformers
scipy
requests
tensorflow